First commited as before_7472_rev1

+++ systemtap/org.eclipse.linuxtools.threadprofiler/src/org/eclipse/linuxtools/internal/threadprofiler/DataPoint.java
@@ -34,7 +34,7 @@ public class DataPoint {
 	 * @return
 	 */
 	public int getX() {
-		return values[0];
+		return (int) values[0];
 	}
 
 	public void setX(int x) {
@@ -46,7 +46,7 @@ public class DataPoint {
 	 * @return
 	 */
 	public int getY() {
-		return values[1];
+		return (int) values[1];
 	}
 
 	public void setY(int y) {
